Struts in place to support pluggable schemas.
NOTE: It bytestream parser in `image.cpp` is not yet ready to support pluggable
schemas but the basic struts for the comms between the framegrabber and the
camera are now there. There is still quite a bit to work on to make this work
properly. It is under active development.