Create an option to initialize a senor with a camera application
At the moment ifm sells two types of O3D3XX devices a so called camera
with a 5 PIN connector and a sensor with a 8 PIN connector. Both devices
are API compatible but the sensor lacks a default application in factory
default mode.

This patch provides a command line option to create and activate a default
camera application.


Signed-off-by: Christian Ege <k4230r6@gmail.com>