Fixed a race condition with the wait for OEM bootet
When using a NTP server it could happen that there has
a time synchronization taken place right in the time
when we are in the loop to wait for the camera being
booted. The loop will be left because the timeout was
triggered. By using a monotonic clock we ensure that the
jump in time will not influence us. This does not solve
issue #122 which deals with ``WaitForFrame`` being not
error proof against time jumps.

This also fixes an initialization issue with the duration
variable.

Signed-off-by: Christian Ege <christian.ege@ifm.com>