Handle spurious wakeups in WaitForFrame
This is a backport from https://github.com/ifm/ifm3d/pull/180/commits/8d2c18abb1a565e73404d6c0b30a8a0457e6f26a

FrameGrabber utilizes a condition variable to synchronize on frame
availability. Condition variables are subject to spurious wakes and
thus some predicate must be provided to ensure the thread actually
unblocked for the correct reason.

In practice spurious wakes are extremely rare, but they do happen. There
happens to be a bug in the Visual C++ 2017 implementation of
std::condition_variable which causes timeouts of small values to end up
looking like spurious wakes. That's how this issue was discovered.
Nonetheless, a spurious wake can and does occur on all platforms, and
can cause the FrameGrabber to serve empty buffers or previous frames.

This change implements usage of the condition "did the pointer to data
in the front buffer change?", indicating the pointer swap did indeed
occur.

There is a race in this implementation: If the thread calling
WaitForFrame executes the read of the pointer after it has been swapped
but before the CV is signaled, the caller will miss the current frame.
This race should be extremely rare (unless polling WaitForFrame in a
tight loop with small timeout, but this is discouraged for general
performance reasons). The solution to avoid this race would be to signal
the CV under lock, but this simply guarantees the race will be lost by
WaitForFrame, missing the current frame. This also potentially
introduces a performance issue as the signaled thread will immediately
block on the mutex held by the signaler. In summary, the consequences of
losing the race are minimal and given the rarity, not worth handling
further.

Closes #122